GFAnalytics/requirements.txt
# Data processing
pandas>=1.3.0
numpy>=1.20.0

# Machine Learning
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Google Cloud Platform
google-cloud-bigquery>=2.30.0
google-cloud-storage>=1.42.0
google-api-python-client>=2.0.0
google-auth>=2.0.0

# Stock data
yfinance>=0.1.70

# Utilities
pyyaml>=6.0
pytz>=2021.1

# Optional: for ICE data (if using)
# icedata>=1.0.0  # Replace with actual package name and version

# Optional: for advanced ML models
# xgboost>=1.4.0
# lightgbm>=3.2.0
# tensorflow>=2.6.0
# keras>=2.6.0 